requirements.txt
# REMARL dependencies
# Core ML
torch>=2.1.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0
sentence-transformers>=2.7.0
numpy>=1.26.0
scipy>=1.13.0

# ── Local LLM (Ollama) ─────────────────────────────────────────────────────
ollama>=0.2.0                  # ollama Python client (optional helper)
langchain>=0.2.0
langchain-ollama>=0.1.3        # ChatOllama for LangChain agents
langchain-community>=0.2.0
langgraph>=0.1.0

# NLP utilities
nltk>=3.8.1
textstat>=0.7.3                # readability metrics for reward
transformers>=4.40.0

# Data and storage
pyyaml>=6.0.1
pandas>=2.2.0
tqdm>=4.66.0

# Terminal UI (Human-in-the-Loop pipeline)
rich>=13.7.0                   # beautiful terminal output + prompts

# Experiment tracking
tensorboard>=2.16.0
wandb>=0.17.0                  # optional but recommended

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0

# Development utilities
black>=24.4.0
isort>=5.13.0
mypy>=1.10.0
